Redo email endpoint with Astro
